+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Enable access to `WakuMessage.timestamp`.
+- Examples (web chat): Use `WakuMessage.timestamp` as unique key for list items.
 
 ## [0.5.0] - 2021-05-21